Initial commit of ProjectTracker II data model
This is a little django application that was developed to explore
alternative data-models for project tracking.  The single largest
difference between previous versions and this version is that all
reports are tracked in a single table rather than different tables for
different reports.  this should make the data model much more dynamic
as additional reporting requirements can be added as needed.

The project also includes the begining of a form that can be used to
create or edit reports.