Added tests for searching functionality.
For the most part, these tests just verify that the testing form can be rendered and submitted.  I was unable to find a way to test faceting in a test environment since the indices would have to rebuild on the Setup method of each test.  If you find a way, simply un-comment the assert contains in each of the tests.