Added test to verify that get_messages_dict works with dis-contiguous messages.
There was a small bug in the original notification code that was caused by referencing message.id instead of message2user.id. The original simple test passed because it was essentially a 1-to-1 relationship.  The message.id and message2users.id where the same.  The new test created a second user and sends him a bunch of fake messages between the first users first message and this last message.  If everything is working as expected, it shouldn't matter how many other messages are sent to other users - you should always get you messages, an only your messages.