Changed Messages2Users to m2mThrough
Modified the relationship between users, messages2users and messages to
be a many-2-many through.  Also changed a number of the field names to
be more descriptive.  Foreign key to message objects are always
'message' (not 'msg') and the content of the message object is 'msgtxt' (again
not 'msg')

mark_as_read() is now a method of a Messages2Users object rather than a
stand alone function.

Most tests in test_messages.py pass.  Two failing tests are associated
with mark_as_read functionality.