Fixed bug associated with protected fields being changed.
There was a bug that resulted in protected milestones being revoked when
non-managers edited a project form.  It was caused by the absence of
protected milestones in the returned list of id - without the id, it
looked like those milestones where revoked.  A simple if statement in
the form checks to see if the user is a manager - in which case
protected milestone id are added to the return list.

A couple of integration tests were written to catch this situation.

All tests pass.