Added views, urls and template code for manager buttons on detail page
Added views to approve, unapprove and sign-off of individual projects.
These are called from the project detail page, but are only available to
managers.  Code was added to detail template to only render those
buttons if the user is manager.

Test to verify this code have not been written yet.