Updates associated with GetTrackerData
The biggest changes in this commit are associated with updates to
migration scripts to retrieve data from the MS Access version of Project
Tracker and insert it directly into Postgres.  The migration no longer
goes through an intermediate sqlite database.

Other changes in this commit were small tweaks to settings
files (changes in ip addresses ect).  Additionally, a large number of
changes were associated with removal of trailing white spaces in any
file that was opened and subsequently saved.

Most of the tests pass.  Three tests associated with file upload
continue to fail.  The failures seem to be associated with the tests as
the functionality continues to work as expected.