Add functionality to link to other projects.
The fields comment_html and risk_risk where added to project model.
and they are updated automatically with the markdown of comment and risk
fields when the project object is saved.

Also added functionality to replace text of the pattern "project:
LHA_XX99_999" with a link to the appropriate project when the pattern is
written in the description or risk text boxes.

New tests were added to ensure that project links work as expected in
both the comment and risk field.

Most tests pass, the three long-standing test failures remain.