Added Field Lead to project model.
Field Lead is now an optional field for any project.  Field Lead can be
any valid user and can be added to project in project detail form.
Field Lead is only rendered on project detail if one is associated with
the project.  Unit tests added to test_view2.py to verify behaviour.