Updates to report upload and message tests
This commit contains a couple of small tests that fix long-standing
problems in the tests.  One fixes an error in report upload tests -
there was a missing space in the 'shouldbe' path.  The other updates the
created timestamp on messages2users so that they are distinct and sort
properly.  Tests associated with uploading associated files continue to
fail.  Its not clear why.