Added salary and odoe to project model including tests
The field total_cost was removed from the project model and replaced
with fields for salary and odoe.  A total_cost() method was added to
project model which is simply the sum of odoe and salary.  Tests to
verify that total_cost() works as expected, verify that form constrains
function properly, and that the cost can be updated and rendered on project detail
template have been included in this commit.