Added script to automatically backup Project Tracker II
The script introduced in this commit is intended to be run as a
scheduled task and backs up all of the reports, associated files, and
data tables in pjtk2.  Reports are copied to a directory structure that
mirrors that of MEDIA_URL. All of the tables in pjtk2 are exported in
pjtk2.sql.  A command line option allows for either complete or
incremental backup of reports.  A small report associated with each
backup (including date and time, number of files found and any errors)
is written to a log file ({trg_dir}/backup.log).