Project object manager returns only active projects
The default model manager for projects was returning all projects,
including those with an active status of False.  A get_query_set()
method was added to the model manager for projects and tests were
updated to ensure that it was returning the expected number of records
given the value of 'active'.