Create User Project List View
A user project list was created to show all of the projects that a
particular user was associated with, either as project lead or field
lead.  The view is most easily accessible from hyperlinks on the project
detail page.  Tests were included in /tests/integration/test_view.py to
verify that the lists rendered as expected depending on role of the user.