Small change to project leader list view
The project list template always included the string "projects
associated with" even when no lead was specified.  Caused by handling of
non-existent project leads (resulted in the same return value as no
project lead).  The get_context() method of the project list view was
modified to account for three possible situations: no project leader
name, a project leader with projects, a project leader that does not
exist (or has not projects).