Added tests for New Project, Project Form and
Added a test to verify that the link to New project was rendered in base
template when a user is logged in.  Additionally, there was a problem
with project form when a new project was being created.  The user filter
was not configured correctly.  Tests were added to verify that it does now.