Changed field type for salary and odoe.
Originally, these fields where integers on the project detail form, but
were pre-populated with 0.00 by default.  This would result in an error
when the form was saved if the values were not changed to an integer
value.  The fields where changed to decimal numbers with a maximum of
two digits.