Added Report Description page and updated tests
The report description page is simply a flat html file ported directly
from the original org-mode project description file. It is likely to be
edited shortly as the reporting requirements are finalized.  A link was also
added to the page under About in the main menu.  Tests were added to
ensure the url worked.

Tests were also updated to accommodate changes to the currency fields on
project form.