Added shared field and method to milestone objects
In order to make the sharing of milestone objects dynamic, a field
'shared' and a shared() method were added to milestone objects.  Now
milestones that are to be shared among sister projects can be retrieved
using Milestone.objects.shared(), and if a shared milestone report is
uploaded using the report upload form, it will automatically be
associated with any sister projects (regardless of its label).