Updated filtering of project lists to carry accross pages
This commit fixes a long-standing bug - django-filter and
django-pagination were not configured to work properly.  When a filter
was applied and it resulted in more than one page, the filter was lost
when the next page was viewed.  That bug has been addressed in this
commit.  Required the addition of django-pagination and a special
pagination template.