Filtered My Projects to only return projects from the last five years.
In order to improve performance, the number of projects returned on the
submitted, approved and completed tabs of the my projects view, has
been limited to the last 5 years.  Without the filter, the lake managers
were seeing data back to 1960's.  After 5-years the reporting
requirements of a project should be satisfied.