Added employee specific page for managers
The default view for My Projects for managers contains all of the
employees under them.  This can make it difficult to see the status of a
project run by a single employee.  The views in this commit make a
filtered list of projects and their associated reporting milestones
available to managers.  The view is accessed through a link on the
project record on MyProjects.