Added cancelled fields to project and associated view and tests.
There has been a request from a manager to add a cancelled flag to
projects. These are projects that were approved, and may have been
planned but were cancelled for some reason.  Fields 'cancelled' and
'cancelled_by' where added to the project model.  A view was created to
cancel projects and the project detail template was modified to handle
logic associated with cancelling projects.  Tests were added to ensure
that only managers can cancel approved projects.