Final changes to implement cancel and milestone.
In order for the notification system to provide a notice to project
leads when a project was cancelled, it had to be created as a
milestone.  Unfortunately, cancelled still needs to be carried as a
field in project table to keep model managers simple (or working at
all).  The cancel_project view was updated to add an appropriate cancel
milestone, make any outstanding requirements optional and update
cancelled and cancelled by fields of pjtk2_project.  Tests were updated
according and all tests pass (the three exceptions that have been
previously noted).