Updates to pjtk2 tests
This commit is associated with changes to files in pjtk2 associated with
upgrade to python3 and django 1.8.  Most tests pass.  Some tests
associated with other applications (i.e. ticket tracker) still fail.  A
couple of tests associated with the defunct olwidget also fail, but will
be replaced in the near future.