Added alerts for abstract length and Proj Remarks
Added alerts to notify logged-in users if there project abstracts are
approaching or exceeding the maximum character limit.  Tests were
included to ensure that only users with edit writes can see the
warnings.  Similarly, the Comments and Remarks section is only available
to users who are logged in as they are considered to be internal.