Added fisheye_button template tag and tests
Created a template tag that will insert a bootstrap styled button linked
to detail page in fisheye, creel portal and fsis-II for those projects
once the data has been merged.  Commit includes unit tests for the tag
to verify that it works as expected.  The commit also includes some unit
tests for a helper function that removes cut-and-paste carriage returns
from project abstract and remarks fields to that paragraph render
properly using latex.