Improvements associated with MyProjects
The performance of the original version of the MyProject view and
associated template/template tags was poor - in most cases it was
issuing an a large number of queries for each project. This version
gathers the information need for each project with 3 queries and then
converts this information into a dictionary for for use in the
template.  Additionally, indexes where placed on all of the
appropriate fields in tables for projects, bookmarks and messages.
Performance appears to be much improved.  All tests currently pass
although test have not been written for new method added to Project
class.  That's next.