Added extrascripts block to all templates
Added an extrascript block to all of the templates with javascript and
moved the javascript to the bottom of the page.  Also updated the
employee milestone report view to use the new _milestone_table
template and removed the old milestone_status template tag.  Add a
next parameter to the links to update reporting requirements so that
they return to the page that they were called from.  All tests
currently pass.