Added Re-open Project methods and tests
Added a way to re-open projects taht have been closed.  Once a Project
is completed, there is not way to edit its content.  The reOpen method
allows projects to be re-activated so that they can be modified or
updated as needed.  Integration tests verify that the views are
rendered properly and include the correct elements.  Most of the
change to associated test files were to add a Sign Off milestone to
each project in order for the tests to render the template properly.
All of the existing tests continue to pass.