Implimented basic image upload and display.
Added basic components for image upload from imadeit.  Added a
ProjectImage model, associated views, forms, and tempalate to upload
images and display them in a carousel on the project detail page.
Basic functionally works. No tests have been written yet, and it is
not possible for users to edit captions, or delete or re-order images.