Finished implementing image upload functionality
Added views to sort, edit and delete images associated with project.
Also added logic to limit the number of images that can be designated
for inclusion in the annual report.  Currently controlled by a global
variable in the views file, but this could be moved into the
settings. Next - write the tests to make sure all of this works as
expected.