Python 3.7 and pipenv
This commit is larger than I originally indented but includes all the
changes need to upgrade to python 3.7, django 1.11.18 and pipenv.
Most of the changes were small tweaks to tests associated with how
passwords are handled in newest versions of django, but also includes
other changes to eliminate depreciation warnings.