Fixed migrations and upgraded to django 2.2.7
Upgraded to django 2.2.7 and updated the migrations so that all of the
tests will continue to pass. Errors were being caused in the
migrations because ProjectPolygon was originally named ProjectPoly -
this renaming was not being captured completely or accurately in the
migrations causing errors - either missing fields or fields already
existing. All existing tests continue to pass.