Moved a bunch of one-time use scripts up to root directory
This commit includes a large of files, but is mostly just a
project/directory re-org.  There were a number of files and associated
scripts in a pjtk2/utils directory. These were moved up to the root
directory. main.py was renamed as runserver.py  since that is what it
does.  All tests continue to pass.