Added status field to Project Model
Status use to be a method on Project, but required several expensive
queries for each instance - renamed the status method to _get_status()
and added a choice field to the Proejct Model. Updated the tests to
ensure that Approved and Signe Off milestones were always present
when needed. All existing tests continue to pass.