Detail map zooms to lake and project list includes select_related
This is a relatively small commit - the map on the detail page for a
map now zooms to the extent of the Ontario waters for the lake that
the project was run in.  A select related clause was also added to the
project list to reduce the number of queries run each time the view
was rendered.  All existing tests continue to pass.