Updates to Find by ROI and improved query performance
Added select_related and prefetch related to several querysets in
order to cut down on the number of queries run to present some list
views. Updated views and functions associated with region of
interests. Added MultiPolygon table with a 1:1 to projects to ensure
that projects with less than three points (and no convex) are
included.  All existing tests continue to pass.