Removed references to mapbox and maki-markers. cleaned urls.
The class mapbox api is no longer supported. This commit includes the
removal of any reference to mapbox and replaces it with openstreet map
and leaflet.  This commit also includes the removal of maki-markers
which are no longer used, but were being loaded in the project detail
page. The urls were cleaned up a little by adding underscores between
distinct words: projectdetail becomes project_detail.  All tests
continue to pass.