Updates to my_projects to reduce query count
The my_projects view, and the employee project views were extremely
slow due to a huge number of queries associated with each project and
associated milestones. This commit includes updates to reduce the
number of queries and format the data in Python before passing it to
the template. New tests should be written to ensure that the helper
functions work as required.  All exsiting tests pass.