Bug fix to address missing Sign Off milestone and DateFormat
Jeremy Holden identitified a couple of bugs that need to be
immediately fixed. One was with the Date format specified by the date
picker in the project page - "yyyy" was producing dates with year
repeated (20202020).  The other was bug was being caused by the
absence of the signoff milestone. If the milestone was absent, the
.complete() method of a project was throwing an exception.  This has
now been addressed. All existing tests continue to pass.