Fixed bugs associated with Project Owners
There were some issues with the logic associated with project
ownership that was only apparent for regular employees.  The field is
required, but not available, or populated for projects created or
copied.  This commit addresses that, and add some tests to verify that
it work. Additionally, the distinction between managers, employees,
and dba used both the django groups, and the employee.role attribute -
this was confusing.  All roles in project tracker are now controlled
by employee.role.  Django groups will be used for project wide
settings. All existing tests pass.