Updated sister projects to include only lake and protocol
Originally, the sister project candisate list included approved
projects of the same type run in the same year.  This was no longer
adequate when we included other lakes and split project type and
protocol into seperate fields. THis commit includes changes to the
Project Molel to ensure that only projects run in the same lake, using
the same protocol are not included in the list of sister candiates.
Updates to the associated tests are also included. All existing tests
continue to pass.