Project funding odoe and salary as 0 if null
There was a bug in the original total method of a project funding
source object. If either odoe or salary where null, it could cause an
error when calculating the total funding.  This commit update the
method to treat missing values as 0 so that the total can be
calculated.  Additional tests where included to ensure that the method
works as expected. All existing tests continue to pass.