Updating Approved Projects View
The original Approve project form was never intended for use across
multiple lakes - the queries were inefficient, and the number of
projects returned is huge without any way to filter them. This update
introduced a new Approve project form that uses much more efficient
queries to both retrieve the project and update their status.  It also
includes some javascript in the template to allow projects to be
filtered by lake, project lead, and/or project type.  Not fully
tested.  Next steps - replace the existing form and views with the new
one and update the tests.