Fixed Project Owner bug for employees
There was a bug in the form and associated view for creating new
projects in project tracker for regular employees. This has been a
source of frustration for several months now. Essentially, employees
are supposed to be the owner of any new projects that they create, but
the owner field was note editable for employees, preventing the
creation of valid projects. This commit includes a clean_owner method
that return the request user if they are not a manager or dba. Tests
have been updated to reflect the new structure. All existing tests
continue to pass.