Added Api endpoint for report abstracts and associated tests
This commit includes an api view, serializers and associated tests to
provide all of the data elements to create the annual UGLMU assessment
report. The data returned by the serializer should be a direct
drop-in replacement for the existing sql queries.  All of the new and
existing tests continue to pass.