Starting to implement Spatial Points Upload
This commit includes basic url, view, form, and tests associated with
a basic upload functionality to associate points with a project. The
points can be either in csv or xlsx format.  All of the tests are
failing stubs to act as placeholders, and the view doesn't actually do
anything right now.