Version 0.0.4a
This is small update to capture a couple of bugs that had been
reported. The spatial points upload wasn't working if the csv file was
created in xlsx - it turns out that Microsoft products often include a
byte-order-marker on the first line of files they create. The marker
is invisible, but there, causing malformed header issue.
Additionally, the associated file upload has issues with encoding in
the file hash. These have now been fixed as work as expected.  All of
the active tests continue to pass.