Version 0.0.4b
This is a bug fix to address issues with canceling and uncanceling
projects. The logic as being duplicated and the project status flag
was not being updated in the associated views. The views now use the
cancel and uncancel methods on project objects. Three additional
integration tests ensure that the milestone where being created
properly and that the attributes of the project remain in sync.  All
existing tests continue to pass.