Improved spatial point upload error message
The spatail point upload had a couple of bugs - if there were cells to
the right of the header row that had been edited in an xlsx file, they
were being parsed as 'None' but appeared empty in the spreadsheet -
raising a malformed header error on headers that looked perfect. List
elements of 'None' are now striped off before the comparison is made.
Additionally, the if the header appears to be malformed, the values
recieved are rendered in the error message. Hopefully makeing the
problem easier to identify and fix.  All of the previously passing tests
continue to pass.